+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobi.m
@@ -192,12 +192,9 @@ __attribute__((constructor)) static void initialize_imageCache() {
                            options:(nullable NSArray<GADAdLoaderOptions *> *)options {
   long long placementId = self.placementId;
   if (placementId == -1) {
-    NSString *errorDesc =
-        [NSString stringWithFormat:@"[InMobi] Error - Placement ID not specified."];
-    NSDictionary<NSString *, NSString *> *errorInfo = @{NSLocalizedDescriptionKey : errorDesc};
-    GADRequestError *error = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                         code:kGADErrorInvalidRequest
-                                                     userInfo:errorInfo];
+    NSError *error = GADMAdapterInMobiErrorWithCodeAndDescription(
+        GADMAdapterInMobiErrorInvalidServerParameters,
+        @"[InMobi] Error - Placement ID not specified.");
     [self.connector adapter:self didFailAd:error];
     return;
   }
@@ -252,19 +249,16 @@ __attribute__((constructor)) static void initialize_imageCache() {
 - (void)requestInterstitialAd {
   long long placementId = self.placementId;
   if (placementId == -1) {
-    NSString *errorDesc =
-        [NSString stringWithFormat:@"[InMobi] Error - Placement ID not specified."];
-    NSDictionary<NSString *, NSString *> *errorInfo = @{NSLocalizedDescriptionKey : errorDesc};
-    GADRequestError *error = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                         code:kGADErrorInvalidRequest
-                                                     userInfo:errorInfo];
+    NSError *error = GADMAdapterInMobiErrorWithCodeAndDescription(
+        GADMAdapterInMobiErrorInvalidServerParameters,
+        @"[InMobi] Error - Placement ID not specified.");
     [self.connector adapter:self didFailAd:error];
     return;
   }
 
   if ([self.connector testMode]) {
-    NSLog(@"[InMobi] Please enter your device ID in the InMobi console to recieve test ads from "
-          @"Inmobi");
+    NSLog(@"[InMobi] Please enter your device ID in the InMobi console to receive test ads from "
+          @"InMobi");
   }
 
   _interstitial = [[IMInterstitial alloc] initWithPlacementId:placementId];
@@ -297,12 +291,9 @@ __attribute__((constructor)) static void initialize_imageCache() {
 - (void)requestBannerWithSize:(GADAdSize)adSize {
   long long placementId = self.placementId;
   if (placementId == -1) {
-    NSString *errorDesc =
-        [NSString stringWithFormat:@"[InMobi] Error - Placement ID not specified."];
-    NSDictionary<NSString *, NSString *> *errorInfo = @{NSLocalizedDescriptionKey : errorDesc};
-    GADRequestError *error = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                         code:kGADErrorInvalidRequest
-                                                     userInfo:errorInfo];
+    NSError *error = GADMAdapterInMobiErrorWithCodeAndDescription(
+        GADMAdapterInMobiErrorInvalidServerParameters,
+        @"[InMobi] Error - Placement ID not specified.");
     [self.connector adapter:self didFailAd:error];
     return;
   }
@@ -314,14 +305,11 @@ __attribute__((constructor)) static void initialize_imageCache() {
 
   CGSize size = GADMAdapterInMobiSupportedAdSizeFromGADAdSize(adSize);
   if (CGSizeEqualToSize(size, CGSizeZero)) {
-    NSString *errorDescription =
+    NSString *description =
         [NSString stringWithFormat:@"Invalid size for InMobi mediation adapter. Size: %@",
                                    NSStringFromGADAdSize(adSize)];
-    NSDictionary<NSString *, NSString *> *errorInfo =
-        @{NSLocalizedDescriptionKey : errorDescription};
-    GADRequestError *error = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                         code:kGADErrorMediationInvalidAdSize
-                                                     userInfo:errorInfo];
+    NSError *error = GADMAdapterInMobiErrorWithCodeAndDescription(
+        GADMAdapterInErrorBannerSizeMismatch, description);
     [self.connector adapter:self didFailAd:error];
     return;
   }
@@ -368,15 +356,7 @@ __attribute__((constructor)) static void initialize_imageCache() {
 }
 
 - (void)banner:(nonnull IMBanner *)banner didFailToLoadWithError:(nonnull IMRequestStatus *)error {
-  NSInteger errorCode = GADMAdapterInMobiAdMobErrorCodeForInMobiCode(error.code);
-  NSDictionary<NSString *, NSString *> *errorInfo =
-      @{NSLocalizedDescriptionKey : error.localizedDescription};
-  GADRequestError *reqError = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                          code:errorCode
-                                                      userInfo:errorInfo];
-  [self.connector adapter:self didFailAd:reqError];
-  NSLog(@"<<<< ad request failed.>>>, error=%@", error);
-  NSLog(@"error code=%ld", (long)error.code);
+  [self.connector adapter:self didFailAd:error];
 }
 
 - (void)banner:(nonnull IMBanner *)banner didInteractWithParams:(nonnull NSDictionary *)params {
@@ -422,15 +402,7 @@ __attribute__((constructor)) static void initialize_imageCache() {
 
 - (void)interstitial:(nonnull IMInterstitial *)interstitial
     didFailToLoadWithError:(IMRequestStatus *)error {
-  NSLog(@"interstitial did fail with error=%@", error.localizedDescription);
-  NSLog(@"error code=%ld", (long)error.code);
-  NSInteger errorCode = GADMAdapterInMobiAdMobErrorCodeForInMobiCode(error.code);
-  NSDictionary<NSString *, NSString *> *errorInfo =
-      @{NSLocalizedDescriptionKey : error.localizedDescription};
-  GADRequestError *reqError = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                          code:errorCode
-                                                      userInfo:errorInfo];
-  [self.connector adapter:self didFailAd:reqError];
+  [self.connector adapter:self didFailAd:error];
 }
 
 - (void)interstitialWillPresent:(nonnull IMInterstitial *)interstitial {
@@ -446,15 +418,7 @@ __attribute__((constructor)) static void initialize_imageCache() {
 
 - (void)interstitial:(nonnull IMInterstitial *)interstitial
     didFailToPresentWithError:(IMRequestStatus *)error {
-  NSLog(@"interstitial did fail with error=%@", error.localizedDescription);
-  NSLog(@"error code=%ld", (long)error.code);
-  NSInteger errorCode = GADMAdapterInMobiAdMobErrorCodeForInMobiCode(error.code);
-  NSDictionary<NSString *, NSString *> *errorInfo =
-      @{NSLocalizedDescriptionKey : error.localizedDescription};
-  GADRequestError *reqError = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                          code:errorCode
-                                                      userInfo:errorInfo];
-  [self.connector adapter:self didFailAd:reqError];
+  [self.connector adapter:self didFailAd:error];
 }
 
 - (void)interstitialWillDismiss:(nonnull IMInterstitial *)interstitial {
@@ -488,14 +452,6 @@ __attribute__((constructor)) static void initialize_imageCache() {
  * Notifies the delegate that the native ad has finished loading
  */
 - (void)nativeDidFinishLoading:(nonnull IMNative *)native {
-  if (_native != native) {
-    GADRequestError *reqError = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                            code:kGADErrorNoFill
-                                                        userInfo:nil];
-    [self.connector adapter:self didFailAd:reqError];
-    return;
-  }
-
   _nativeAd =
       [[InMobiMediatedUnifiedNativeAd alloc] initWithInMobiUnifiedNativeAd:native
                                                                    adapter:self
@@ -508,14 +464,7 @@ __attribute__((constructor)) static void initialize_imageCache() {
  */
 - (void)native:(nonnull IMNative *)native didFailToLoadWithError:(nonnull IMRequestStatus *)error {
   NSLog(@"Native Ad failed to load");
-  NSInteger errorCode = GADMAdapterInMobiAdMobErrorCodeForInMobiCode(error.code);
-  NSDictionary<NSString *, NSString *> *errorInfo =
-      @{NSLocalizedDescriptionKey : error.localizedDescription};
-  GADRequestError *reqError = [GADRequestError errorWithDomain:kGADMAdapterInMobiErrorDomain
-                                                          code:errorCode
-                                                      userInfo:errorInfo];
-
-  [self.connector adapter:self didFailAd:reqError];
+  [self.connector adapter:self didFailAd:error];
 }
 
 /**