+++ adapters/InMobi/InMobiAdapter/GADMediationAdapterInMobi.h
@@ -16,6 +16,15 @@
 #import <GoogleMobileAds/GoogleMobileAds.h>
 #import <InMobiSDK/InMobiSDK.h>
 
+typedef NS_ENUM(NSInteger, GADMAdapterInMobiErrorCode) {
+  /// Missing server parameters.
+  GADMAdapterInMobiErrorInvalidServerParameters = 101,
+  /// Banner Size Mismatch.
+  GADMAdapterInErrorBannerSizeMismatch = 102,
+  /// An ad is already loaded for this network configuration.
+  GADMAdapterInMobiErrorAdAlreadyLoaded = 103
+};
+
 @interface GADMediationAdapterInMobi : NSObject <GADMediationAdapter>
 
 @end